pyproject.toml
[tool.poetry]
name = "socketio_testing"
version = "0.1.0"
description = ""
authors = ["Mikail Khan <mikail.khan45@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-socketio = "^5.0.1"
Flask = "^1.1.2"
eventlet = "^0.30.2"
requests = "^2.25.1"
websocket-client = "^0.58.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
